rename monitor & move calculator
